evtGAN/requirements.txt
torch
numpy
xarray
pandas
scipy
pyyaml